code refactor, add splash screen and icon, fix imports
