+++ .eslintignore
@@ -1,1 +1,3 @@
 node_modules/
+/yarn.lock
+/package-lock.json
