+++ .gitignore
@@ -11,9 +11,9 @@ npm-debug.*
 web-build/
 .xml
 .lock
+.idea
 
 # macOS
 .DS_Store
 
 /envoirment
-.env
