+++ src/components/FavoriteMovieCard.tsx
@@ -1,9 +1,9 @@
 import { THEME } from '../styles/theme';
-import { Image, Pressable, StyleSheet, Text, View, Dimensions } from 'react-native';
-import { textTranslate } from '../utils/textTranslate';
+import { Image, Pressable, StyleSheet, Text, View } from 'react-native';
+import { textTranslate, GetPosterPath } from '../utils';
 import { NowPlayingResultsType, useAppNavigation } from '../types/types';
 import { useAppContext } from '../context/AppContext';
-import { GetPosterPath } from '../utils/getPosterPath';
+import { DEVICE_WIDTH } from '../constants';
 
 type FavoriteCardType = {
   setParentPage: (page: string) => void;
@@ -48,7 +48,7 @@ export const FavoriteMovieCard = ({ setParentPage, item }: FavoriteCardType) => 
 
 const styles = StyleSheet.create({
   favoriteMovieCardWrapper: {
-    width: Dimensions.get('screen').width,
+    width: DEVICE_WIDTH,
     backgroundColor: THEME.DARK,
     marginVertical: 5,
     flexDirection: 'row',
@@ -60,7 +60,7 @@ const styles = StyleSheet.create({
   text: {
     color: THEME.TEXT,
     fontSize: 20,
-    width: Dimensions.get('screen').width - 120,
+    width: DEVICE_WIDTH - 120,
   },
   left: {
     flexDirection: 'row',
