+++ src/constants/constants.ts
@@ -1,3 +1,8 @@
+import { Dimensions } from 'react-native';
+
+export const api_key = '365da808b8c3c4b7a04da7eca0d9524e';
+export const BASE_URL = 'https://api.themoviedb.org/3/';
+
 export const CONSTANTS = {
   ICON_SIZE: 34,
   TEXT10: 10,
@@ -10,3 +15,5 @@ export const ERROR_MESSAGES = {
   REQUIRE_FIELD: 'Field is require',
   INCORRECT_LOGIN_OR_PASSWORD: 'Incorrect Login or Password',
 };
+
+export const DEVICE_WIDTH = Dimensions.get('window').width;
