+++ src/screens/nestedScreen/index.ts
@@ -3,3 +3,4 @@ export * from './MainNested';
 export * from './NowPlayingMovies';
 export * from './FavoriteList';
 export * from './SearchMoviesNested';
+export * from './SearchMovies';
