+++ src/utils/index.ts
@@ -1,0 +1,3 @@
+export * from './textTranslate';
+export * from './async-store';
+export * from './getPosterPath';
